simplify and cleanup
