+++ plugins/gcp/fix_plugin_gcp/collector.py
@@ -1,6 +1,6 @@
-from datetime import datetime, timedelta, timezone
 import logging
 from concurrent.futures import ThreadPoolExecutor
+from datetime import datetime, timezone
 from typing import Type, List, Any, Optional, cast
 
 from fix_plugin_gcp.config import GcpConfig
@@ -161,30 +161,19 @@ class GcpProjectCollector:
 
     def collect_usage_metrics(self, builder: GraphBuilder) -> None:
         for resource in builder.graph.nodes:
-            if not isinstance(resource, GcpResource):
-                continue
-            # region can be overridden in the query
-            if region := cast(GcpRegion, resource.region()):
-                resource_queries: List[monitoring.GcpMonitoringQuery] = resource.collect_usage_metrics(builder)
-                if resource_queries:
-                    start = builder.metrics_delta
+            if isinstance(resource, GcpResource) and (mq := resource.collect_usage_metrics(builder)):
 
-                    def collect_and_set_metrics(
-                        start: timedelta,
-                        region: GcpRegion,
-                        queries: List[monitoring.GcpMonitoringQuery],
-                        resource: GcpResource,
-                    ) -> None:
-                        start_at = builder.created_at - start
-                        try:
-                            result = monitoring.GcpMonitoringMetricData.query_for(
-                                builder.for_region(region), queries, start_at, builder.created_at
-                            )
-                            monitoring.update_resource_metrics(resource, result)
-                        except Exception as e:
-                            log.warning(f"Error occurred in region {region}: {e}")
+                def collect_and_set_metrics() -> None:
+                    start_at = builder.created_at - builder.metrics_delta
+                    region = cast(GcpRegion, resource.region())
+                    try:
+                        rb = builder.for_region(region)
+                        result = monitoring.GcpMonitoringMetricData.query_for(rb, mq, start_at, builder.created_at)
+                        monitoring.update_resource_metrics(resource, result)
+                    except Exception as e:
+                        log.warning(f"Error occurred in region {region}: {e}")
 
-                    builder.submit_work(collect_and_set_metrics, start, region, resource_queries, resource)
+                builder.submit_work(collect_and_set_metrics)
 
     def remove_unconnected_nodes(self, builder: GraphBuilder) -> None:
         def rm_leaf_nodes(clazz: Any, ignore_kinds: Optional[Type[Any]] = None) -> None:
